# libs/domains/agenstra/backend/feature-agent-controller/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Controller API
  version: 1.0.0
  description: HTTP API for managing clients and proxied agent operations
  contact:
    name: ForePath
    email: hi@forepath.io
    url: https://forepath.io
servers:
  - url: http://localhost:3100/api
    description: Local development
  - url: http://localhost:3100
    description: Root (OpenTelemetry metrics; outside /api prefix)
security:
  - bearerAuth: []
paths:
...
  /clients:
    get:
      summary: List clients
      description: |
        Returns all clients the user has access to. In api-key mode, returns all clients.
        For regular users, only returns clients where:
        - The user is the creator (user_id matches)
        - The user is in the client_users relationship table
        Global admins see all clients.
      operationId: listClients
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 0
          description: Maximum number of clients to return (default 10)
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
          description: Number of clients to skip (default 0)
        - in: query
          name: search
          schema:
            type: string
          description: Case-insensitive search across name, description, id
      responses:
        '200':
          description: Array of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientResponseDto'
    post:
      summary: Create client
      description: |
        Creates a new client. The user_id is automatically set to the logged-in user.
        In api-key mode, user_id is set to null.
      operationId: createClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientDto'
      responses:
        '201':
          description: Created client with generated API key (if applicable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateClientResponseDto'
  /clients/{id}:
    get:
      summary: Get client by id
      description: |
        Returns a client by ID. Only accessible if the user has access to the client.
        Access is granted if:
        - The user is a global admin
        - The user is the creator (user_id matches)
        - The user is in the client_users relationship table
        In api-key mode, access is always granted.
      operationId: getClient
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponseDto'
        '403':
          description: User does not have access to this client
    post:
      summary: Update client
      description: |
        Updates a client. Requires workspace management rights: global admin, workspace creator,
        client_users role `admin`, API key authentication, or api-key server mode (not plain client_users `user`).
        See `canManageWorkspaceConfiguration` on ClientResponseDto.
      operationId: updateClient
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientDto'
      responses:
        '200':
          description: Updated client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponseDto'
        '403':
          description: |
            Forbidden — no access to this client, or caller lacks workspace management rights
            (same rule as POST for updates).
    delete:
      summary: Delete client
      description: |
        Deletes a client. Same workspace management requirement as POST /clients/{id} (not available to plain client_users `user`).
      operationId: deleteClient
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '403':
          description: |
            Forbidden — no access to this client, or caller lacks workspace management rights.
...
  /clients/{id}/agents/{agentId}/files:
    get:
      summary: List directory contents (proxied)
      operationId: listClientAgentDirectory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
        - in: query
          name: path
          schema:
            type: string
          description: Directory path relative to the selected context root (defaults to '.')
        - $ref: '#/components/parameters/fileManagerContext'
      responses:
        '200':
          description: Array of file nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileNodeDto'
        '400':
          description: Invalid context or agent type does not support configuration file access (proxied from agent-manager)
        '403':
          description: User does not have access to this client, or lacks workspace management rights when context is config
        '404':
          description: Client, agent, or directory not found
...
  /clients/{id}/agents/{agentId}/files/{path}:
    head:
      summary: Probe file metadata without transferring the body (proxied)
      operationId: headClientAgentFile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
        - in: path
          name: path
          required: true
          schema:
            type: string
          description: File path relative to the selected context root (supports nested paths)
        - $ref: '#/components/parameters/fileManagerContext'
      responses:
        '200':
          description: Metadata headers only (no body)
          headers:
            Accept-Ranges:
              schema: { type: string, enum: [bytes] }
            X-File-Type:
              schema:
                $ref: '#/components/schemas/AgentFileType'
            Content-Type:
              schema: { type: string }
            Content-Length:
              schema: { type: integer }
        '400':
          description: Invalid path, invalid context, or missing access
        '404':
          description: Client, agent, or file not found
    get:
      summary: Read file content as raw bytes (proxied)
      operationId: readClientAgentFile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
        - in: path
          name: path
          required: true
          schema:
            type: string
          description: File path relative to the selected context root (supports nested paths)
        - $ref: '#/components/parameters/fileManagerContext'
        - in: query
          name: download
          schema:
            type: boolean
          description: When true, set Content-Disposition attachment with the file basename
        - in: header
          name: Range
          schema:
            type: string
          description: Optional bytes range (e.g. bytes=0-1023) for partial content
      responses:
        '200':
          description: Full file content
          headers:
            Accept-Ranges:
              schema: { type: string, enum: [bytes] }
            X-File-Type:
              schema:
                $ref: '#/components/schemas/AgentFileType'
            Content-Disposition:
              schema: { type: string }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
            application/pdf:
              schema: { type: string, format: binary }
            text/plain:
              schema: { type: string, format: binary }
            image/*:
              schema: { type: string, format: binary }
            video/*:
              schema: { type: string, format: binary }
            audio/*:
              schema: { type: string, format: binary }
        '206':
          description: Partial content for Range requests
          headers:
            Accept-Ranges:
              schema: { type: string, enum: [bytes] }
            Content-Range:
              schema: { type: string }
            X-File-Type:
              schema:
                $ref: '#/components/schemas/AgentFileType'
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
            application/pdf:
              schema: { type: string, format: binary }
            text/plain:
              schema: { type: string, format: binary }
            image/*:
              schema: { type: string, format: binary }
            video/*:
              schema: { type: string, format: binary }
            audio/*:
              schema: { type: string, format: binary }
        '400':
          description: Invalid path, invalid context, or agent type does not support configuration file access
        '403':
          description: User does not have access to this client, or lacks workspace management rights when context is config
        '404':
          description: Client, agent, or file not found
        '416':
          description: Range not satisfiable
    put:
      summary: Write file content as raw bytes (proxied; full file or Content-Range chunk)
      operationId: writeClientAgentFile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
        - in: path
          name: path
          required: true
          schema:
            type: string
          description: File path relative to the selected context root (supports nested paths)
        - $ref: '#/components/parameters/fileManagerContext'
        - in: header
          name: Content-Range
          schema:
            type: string
          description: Optional chunk range (bytes start-end/total) for chunked uploads
        - in: header
          name: X-Upload-Id
          schema:
            type: string
            format: uuid
          description: Optional upload session id for chunked uploads
        - in: header
          name: X-File-Type
          schema:
            $ref: '#/components/schemas/AgentFileType'
          description: Optional client classification hint
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
          application/pdf:
            schema: { type: string, format: binary }
          text/plain:
            schema: { type: string, format: binary }
          image/*:
            schema: { type: string, format: binary }
          video/*:
            schema: { type: string, format: binary }
          audio/*:
            schema: { type: string, format: binary }
      responses:
        '204':
          description: File written successfully (or chunk accepted; last chunk finalizes)
        '400':
          description: Invalid path, content too large, invalid range, invalid context, or agent type does not support configuration file access
        '403':
          description: User does not have access to this client, or lacks workspace management rights when context is config
        '404':
          description: Client or agent not found
    delete:
      summary: Delete file or directory (proxied)
      operationId: deleteClientAgentFileOrDirectory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
        - in: path
          name: path
          required: true
          schema:
            type: string
          description: File or directory path relative to the selected context root (supports nested paths)
        - $ref: '#/components/parameters/fileManagerContext'
      responses:
        '204':
          description: File or directory deleted successfully
        '400':
          description: Invalid path, invalid context, or agent type does not support configuration file access
        '403':
          description: User does not have access to this client, or lacks workspace management rights when context is config
        '404':
          description: Client, agent, or file/directory not found
    patch:
      summary: Move file or directory (proxied)
      operationId: moveClientAgentFileOrDirectory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
        - in: path
          name: path
          required: true
          schema:
            type: string
          description: Source file or directory path relative to the selected context root (supports nested paths)
        - $ref: '#/components/parameters/fileManagerContext'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveFileDto'
      responses:
        '204':
          description: File or directory moved successfully
        '400':
          description: Invalid path, destination, invalid context, or agent type does not support configuration file access
        '403':
          description: User does not have access to this client, or lacks workspace management rights when context is config
        '404':
          description: Client, agent, or source file/directory not found
    post:
      summary: Create file or directory (proxied)
      operationId: createClientAgentFileOrDirectory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the client
        - in: path
          name: agentId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the agent
        - in: path
          name: path
          required: true
          schema:
            type: string
          description: File or directory path relative to the selected context root (supports nested paths)
        - $ref: '#/components/parameters/fileManagerContext'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFileDto'
      responses:
        '201':
          description: File or directory created
        '400':
          description: Invalid request, invalid context, or agent type does not support configuration file access
        '403':
          description: User does not have access to this client, or lacks workspace management rights when context is config
        '404':
          description: Client or agent not found
...
  /admin/opencode/config/files:
    get:
      tags: [Admin]
      summary: List children under a global layer path
      operationId: listGlobalOpencodeLayerFiles
      parameters:
        - in: query
          name: path
          schema:
            type: string
            default: '.'
      responses:
        '200':
          description: Directory listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeLayerFileListResponseDto'
    post:
      tags: [Admin]
      summary: Create a global layer file or directory
      operationId: createGlobalOpencodeLayerFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOpencodeLayerFileDto'
      responses:
        '200':
          description: Created entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeLayerFileResponseDto'
  /admin/opencode/config/files/{path}:
    get:
      tags: [Admin]
      summary: Get a global OpenCode layer file
      operationId: getGlobalOpencodeLayerFile
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
          description: Path as entered in config (no managed prefix rewrite)
      responses:
        '200':
          description: Global layer file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeLayerFileResponseDto'
        '404':
          description: Layer file not found
    put:
      tags: [Admin]
      summary: Create or update a global OpenCode layer file
      operationId: putGlobalOpencodeLayerFile
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertOpencodeLayerFileDto'
      responses:
        '200':
          description: Saved global layer file (fan-out emit started)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeLayerFileResponseDto'
    delete:
      tags: [Admin]
      summary: Delete a global OpenCode layer file
      operationId: deleteGlobalOpencodeLayerFile
      parameters:
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
...
  /clients/{id}/opencode/config/files:
    get:
      summary: List children under a workspace layer path
      operationId: listClientOpencodeLayerFiles
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: path
          schema:
            type: string
            default: '.'
      responses:
        '200':
          description: Directory listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeLayerFileListResponseDto'
    post:
      summary: Create a workspace layer file or directory
      operationId: createClientOpencodeLayerFile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOpencodeLayerFileDto'
      responses:
        '200':
          description: Created entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeLayerFileResponseDto'
  /clients/{id}/opencode/config/files/{path}:
    get:
      summary: Get a workspace OpenCode layer file
      operationId: getClientOpencodeLayerFile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: path
          required: true
          schema:
            type: string
          description: Path as entered in config (no managed prefix rewrite)
      responses:
        '200':
          description: Workspace layer file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeLayerFileResponseDto'
        '404':
          description: Layer file not found
    put:
      summary: Create or update a workspace OpenCode layer file
      operationId: putClientOpencodeLayerFile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertOpencodeLayerFileDto'
      responses:
        '200':
          description: Saved workspace layer file (fan-out emit started)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpencodeLayerFileResponseDto'
    delete:
      summary: Delete a workspace OpenCode layer file
      operationId: deleteClientOpencodeLayerFile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  parameters:
    fileManagerContext:
      name: context
      in: query
      required: false
      schema:
        type: string
        enum: [app]
        default: app
      description: Proxied to agent-manager. `app` (default) is the workspace root under `/app`.
  schemas:
    AuthenticationType:
      type: string
      enum:
        - api_key
        - keycloak
...
    ClientResponseDto:
      type: object
      required:
        [
          id,
          name,
          endpoint,
          authenticationType,
          isAutoProvisioned,
          canManageWorkspaceConfiguration,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        endpoint:
          type: string
          format: uri
        authenticationType:
          $ref: '#/components/schemas/AuthenticationType'
        agentWsPort:
          type: integer
          minimum: 1
          maximum: 65535
          description: Optional Socket.IO port for the client's agents gateway
        config:
          $ref: '#/components/schemas/ConfigResponseDto'
          description: Configuration from the agent-manager instance (optional, may be undefined if agent-manager is unreachable)
        isAutoProvisioned:
          type: boolean
          description: Indicates whether the client was auto-provisioned via server provisioning. Used to determine if server info endpoints should be called.
        canManageWorkspaceConfiguration:
          type: boolean
          description: |
            True if the current caller may change prototype autonomy, agent environment variables, agent CRUD,
            and workspace (client) update/delete. Requires global admin, workspace creator, client_users admin,
            API key authentication, or api-key server mode.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateClientDto:
      type: object
      required: [name, endpoint, authenticationType]
      properties:
        name:
          type: string
        description:
          type: string
        endpoint:
          type: string
          format: uri
        authenticationType:
          $ref: '#/components/schemas/AuthenticationType'
        apiKey:
          type: string
          description: API key (optional, will be auto-generated if not provided for API_KEY type)
        keycloakClientId:
          type: string
          description: Keycloak client ID (required for KEYCLOAK authentication type)
        keycloakClientSecret:
          type: string
          description: Keycloak client secret (required for KEYCLOAK authentication type)
        keycloakRealm:
          type: string
          description: Keycloak realm (optional, uses environment default if not provided)
        agentWsPort:
          type: integer
          minimum: 1
          maximum: 65535
          description: Optional Socket.IO port for the client's agents gateway (defaults to 8080 if not set)
    UpdateClientDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        endpoint:
          type: string
          format: uri
        authenticationType:
          $ref: '#/components/schemas/AuthenticationType'
        apiKey:
          type: string
          description: API key (can be updated but will never be included in responses)
        keycloakClientId:
          type: string
          description: Keycloak client ID
        keycloakClientSecret:
          type: string
          description: Keycloak client secret
        keycloakRealm:
          type: string
          description: Keycloak realm
        agentWsPort:
          type: integer
          minimum: 1
          maximum: 65535
          description: Optional Socket.IO port for the client's agents gateway
    CreateClientResponseDto:
      allOf:
        - $ref: '#/components/schemas/ClientResponseDto'
        - type: object
          properties:
            apiKey:
              type: string
              description: The API key for the client (only for API_KEY authentication type, only returned once during creation)
...
    AgentTypeInfo:
      type: object
      required: [type, displayName, capabilities]
      properties:
        type:
          type: string
          description: The unique type identifier (e.g., 'opencode')
        displayName:
          type: string
          description: Human-readable display name (e.g., 'OpenCode')
        capabilities:
          $ref: '#/components/schemas/AgentTypeCapabilities'
...
    ConfigResponseDto:
      type: object
      required: [agentTypes]
      properties:
        gitRepositoryUrl:
          type: [string, 'null']
          description: Git repository URL used for cloning repositories when creating agents
        gitRepositorySetupMode:
          type: string
          enum: [clone, empty]
          description: Default setup mode for new agents on this instance
        agentTypes:
          type: array
          items:
            $ref: '#/components/schemas/AgentTypeInfo'
          description: Array of available agent provider types with display names
    AgentFileType:
      type: string
      enum: [text, binary, pdf, image, video, audio]
      description: Server-authoritative file classification for editor/preview behavior
    FileNodeDto:
      type: object
      required: [name, type, path]
      properties:
        name:
          type: string
          description: File or directory name
        type:
          type: string
          enum: [file, directory]
          description: Node type
        path:
          type: string
          description: Relative path from /app
        size:
          type: [integer, 'null']
          description: File size in bytes (only for files)
        modifiedAt:
          type: [string, 'null']
          format: date-time
          description: Last modification timestamp
    CreateFileDto:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [file, directory]
          description: Type of node to create (empty file or directory; write content via PUT)
    MoveFileDto:
      type: object
      required: [destination]
      properties:
        destination:
          type: string
          description: Destination path relative to /app (supports nested paths)
...
    OpencodeLayerFileResponseDto:
      type: object
      required:
        [
          id,
          scope,
          path,
          emittedPath,
          entryKind,
          content,
          contentSha,
          updatedAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        scope:
          type: string
          enum: [global, workspace]
        clientId:
          type: string
          format: uuid
          nullable: true
        path:
          type: string
          description: Path as entered (sanitized; no managed prefix)
        emittedPath:
          type: string
          description: Same as path; written into agent `/app` trees as entered
        entryKind:
          type: string
          enum: [file, directory]
        content:
          type: string
          description: Plaintext after AES-256-GCM decrypt (empty for directories)
        contentSha:
          type: string
        updatedAt:
          type: string
          format: date-time
    OpencodeLayerFileListEntryDto:
      type: object
      required: [name, path, entryKind]
      properties:
        name:
          type: string
        path:
          type: string
        entryKind:
          type: string
          enum: [file, directory]
    OpencodeLayerFileListResponseDto:
      type: object
      required: [path, entries]
      properties:
        path:
          type: string
        entries:
          type: array
          items:
            $ref: '#/components/schemas/OpencodeLayerFileListEntryDto'
    CreateOpencodeLayerFileDto:
      type: object
      required: [path, entryKind]
      properties:
        path:
          type: string
        entryKind:
          type: string
          enum: [file, directory]
        content:
          type: string
    UpsertOpencodeLayerFileDto:
      type: object
      properties:
        content:
          type: string
        entryKind:
          type: string
          enum: [file, directory]
...